[WRAPPING] Fixed some (very) old issue with SDL2 RWops wrapping

+++ src/include/sdl2rwops.h
@@ -25,8 +25,8 @@ void RWNativeEnd2(SDL2_RWops_t* ops);                           // put emulated 
 int isRWops(SDL2_RWops_t* ops); // 1 if ops seems to be a valid RWops, 0 if not
 
 int64_t RWNativeSeek2(SDL2_RWops_t *ops, int64_t offset, int32_t whence);
-uint32_t RWNativeRead2(SDL2_RWops_t* ops, void* ptr, uint32_t size, uint32_t maxnum);
-int32_t RWNativeWrite2(SDL2_RWops_t *ops, const void *ptr, int32_t size, int32_t num);
+size_t RWNativeRead2(SDL2_RWops_t* ops, void* ptr, size_t size, size_t maxnum);
+size_t RWNativeWrite2(SDL2_RWops_t *ops, const void *ptr, size_t size, size_t num);
 int32_t RWNativeClose2(SDL2_RWops_t* ops);
 
 #endif
